requirements.txt
requests
types-requests
fastapi[standard]
fastapi-restful
fastapi-utils
httpx
typing-inspect
pyzmq
